config-loader: move schema/ to top-level
