+++ packages/config-loader/src/index.ts
@@ -20,14 +20,14 @@
  * @packageDocumentation
  */
 
-export { readEnvConfig, loadConfigSchema, mergeConfigSchemas } from './lib';
+export { loadConfigSchema, mergeConfigSchemas } from './schema';
 export type {
   ConfigSchema,
   ConfigSchemaProcessingOptions,
   ConfigVisibility,
   LoadConfigSchemaOptions,
   TransformFunc,
-} from './lib';
+} from './schema';
 export { loadConfig } from './loader';
 export type {
   ConfigTarget,
