+++ packages/config-loader/src/schema/compile.test.ts
